+++ packages/react-core/src/components/FormSelect/FormSelectOptionGroup.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { Omit } from '../../helpers/typeUtils';
 
 export interface FormSelectOptionGroupProps extends Omit<React.HTMLProps<HTMLOptGroupElement>, 'disabled'> {
   /** content rendered inside the Select Option Group */