+++ packages/react-core/src/components/Nav/NavItem.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import styles from '@patternfly/react-styles/css/components/Nav/nav';
 import { css } from '@patternfly/react-styles';
-import { Omit } from '../../helpers/typeUtils';
 import { NavContext, NavSelectClickHandler } from './Nav';
 
 export interface NavItemProps extends Omit<React.HTMLProps<HTMLAnchorElement>, 'onClick'> {