+++ packages/react-core/src/components/Title/Title.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import { css, getModifier } from '@patternfly/react-styles';
-import { Omit } from '../../helpers/typeUtils';
 import styles from '@patternfly/react-styles/css/components/Title/title';
 import { BaseSizes } from '../../styles/sizes';