+++ packages/react-table/src/components/Table/Table.tsx
@@ -7,7 +7,6 @@ import {
   DropdownPosition
 } from '@patternfly/react-core/dist/js/components/Dropdown/dropdownConstants';
 import { DropdownItemProps } from '@patternfly/react-core/dist/js/components/Dropdown/DropdownItem';
-import { Omit } from '@patternfly/react-core/dist/js/helpers/typeUtils';
 import inlineStyles from '@patternfly/react-styles/css/components/InlineEdit/inline-edit';
 import { css, getModifier } from '@patternfly/react-styles';
 import { Provider } from './base';